Added isStream yes | no option to file scan. isStream is by default no.
isStream should be yes if a file is formatted with niagara:stream
format. In this case it will be parsed with saxdom as a stream - this
is used to use a file to fake a stream.

If isStream is no, then the file is assumed to contain one document
and is parsed with xerces.