Null support.
Also, make result of a group operator a content var for aggregates 
(this is not a change), but it is an element_var for nest! This makes
much more sense. Wow, took me a while to figure out why 
the result of nest was a content_var (which causes
unexpected results.)