Add punctuation support for hashjoin and unnest.
A punctuation has the name space "http://www.cse.ogi.edu/dot/niagara/punct".

Unnest processes punctuation as it does for normal tuple.

Join takes an additional parameter, punctattr, for the punctuating attribute of the input streams, e.g., punctattr="time0 time1", where time0 and time1 are a data attribute of the left and right input stream of the join. Now it only handles linear punctuation and equi-join predicate on the punctuating attributes.